Set a link to each page.
