+++ src/components/layouts/Layout.jsx
@@ -1,21 +1,16 @@
+import React from "react";
+import Navbar from "./Navbar";
+import Container from "@mui/material/Container";
+import Box from "@mui/material/Box";
 
-import React from 'react';
-import Navbar from './Navbar';
-import Container from '@mui/material/Container';
-import Box from '@mui/material/Box';
-
-export default function Layout ({ children }) {
+export default function Layout({ children }) {
   return (
     <>
       <Navbar />
-      <Container component="main" maxWidth="xl">
-        <Box sx={{ py: 4 }}>
-          {children}
-        </Box>
+      <Container component="main" maxWidth="xl" sx={{ mt: 3 }}>
+        <Box sx={{ py: 4 }}>{children}</Box>
       </Container>
       {/* Footer can be added in future tickets */}
     </>
   );
-};
-
-
+}
