+++ src/components/layouts/Navbar.jsx
@@ -1,13 +1,187 @@
-import React from 'react';
+import * as React from "react";
+import AppBar from "@mui/material/AppBar";
+import Box from "@mui/material/Box";
+import Toolbar from "@mui/material/Toolbar";
+import IconButton from "@mui/material/IconButton";
+import Typography from "@mui/material/Typography";
+import Menu from "@mui/material/Menu";
+import MenuIcon from "@mui/icons-material/Menu";
+import Container from "@mui/material/Container";
+import Avatar from "@mui/material/Avatar";
+import Button from "@mui/material/Button";
+import Tooltip from "@mui/material/Tooltip";
+import MenuItem from "@mui/material/MenuItem";
+import AdbIcon from "@mui/icons-material/Adb";
+import { Link } from "react-router-dom";
 
-export default function Navbar  ()  {
+const pages = ["gallery", "best of artwork", "about", "challenge prompt"];
+const settings = ["Logout"];
+function Navbar() {
+  const [isAuth, setIsAuth] = React.useState(false);
+  const [anchorElNav, setAnchorElNav] = React.useState(null);
+  const [anchorElUser, setAnchorElUser] = React.useState(null);
+
+  const handleAuth = () => {
+    setIsAuth(true);
+    setAnchorElNav(null);
+  };
+
+  const handleLogout = () => {
+    setIsAuth(false);
+    setAnchorElUser(null);
+  };
+
+  const handleOpenNavMenu = (event) => {
+    setAnchorElNav(event.currentTarget);
+  };
+
+  const handleOpenUserMenu = (event) => {
+    setAnchorElUser(event.currentTarget);
+  };
+
+  const handleCloseUserMenu = () => {
+    setAnchorElUser(null);
+  };
+
+  const handleCloseNavMenu = () => {
+    setAnchorElNav(null);
+  };
+
   return (
-    <header>
-      {/* Navbar implementation will go here in future tickets */}
-      <div>
-        Navbar Placeholder
-      </div>
-    </header>
-  );
-};
+    <AppBar>
+      <Toolbar disableGutters maxWidth="xl" sx={{ mx: 3 }}>
+        <AdbIcon sx={{ display: "flex", mr: 1, order: { xs: 1, md: 0 } }} />
+        <Typography
+          variant="h6"
+          noWrap
+          component="a"
+          href="/"
+          sx={{
+            mr: 2,
+            display: { xs: "flex", md: "flex" },
+            flexGrow: { xs: 1, md: 0 },
+            order: { xs: 2, md: 0 },
+            fontFamily: "monospace",
+            fontWeight: 700,
+            letterSpacing: ".3rem",
+            color: "inherit",
+            textDecoration: "none",
+          }}
+        >
+          ARTHIVE
+        </Typography>
 
+        <Box sx={{ flexGrow: 1, display: { xs: "flex", md: "none" } }}>
+          <IconButton
+            size="large"
+            aria-label="account of current user"
+            aria-controls="menu-appbar"
+            aria-haspopup="true"
+            onClick={handleOpenNavMenu}
+            color="inherit"
+          >
+            <MenuIcon />
+          </IconButton>
+          <Menu
+            id="menu-appbar"
+            anchorEl={anchorElNav}
+            anchorOrigin={{
+              vertical: "bottom",
+              horizontal: "left",
+            }}
+            keepMounted
+            transformOrigin={{
+              vertical: "top",
+              horizontal: "left",
+            }}
+            open={Boolean(anchorElNav)}
+            onClose={handleCloseNavMenu}
+            sx={{ display: { xs: "block", md: "none" } }}
+          >
+            {pages.slice(0, isAuth ? 4 : 3).map((page) => (
+              <MenuItem key={page} onClick={handleCloseNavMenu}>
+                <Typography
+                  color="primary"
+                  component={Link}
+                  to={page.replaceAll(" ", "-")}
+                  sx={{
+                    textAlign: "center",
+                    textDecoration: "none",
+                    textTransform: "uppercase",
+                  }}
+                >
+                  {page}
+                </Typography>
+              </MenuItem>
+            ))}
+          </Menu>
+        </Box>
+        <Box sx={{ flexGrow: 1, display: { xs: "none", md: "flex" } }}>
+          {pages.slice(0, isAuth ? 4 : 3).map((page) => (
+            <Button
+              component={Link}
+              to={page.replaceAll(" ", "-")}
+              key={page}
+              onClick={handleCloseNavMenu}
+              sx={{ my: 2, color: "white", display: "block" }}
+            >
+              {page}
+            </Button>
+          ))}
+        </Box>
+        {isAuth ? (
+          <>
+            <Tooltip title="Open settings">
+              <IconButton
+                onClick={handleOpenUserMenu}
+                sx={{ p: 0, order: { xs: 3, md: 0 } }}
+              >
+                <Avatar alt="Remy Sharp" src="/static/images/avatar/2.jpg" />
+              </IconButton>
+            </Tooltip>
+            <Menu
+              sx={{ mt: "45px" }}
+              id="menu-appbar"
+              anchorEl={anchorElUser}
+              anchorOrigin={{
+                vertical: "top",
+                horizontal: "right",
+              }}
+              keepMounted
+              transformOrigin={{
+                vertical: "top",
+                horizontal: "right",
+              }}
+              open={Boolean(anchorElUser)}
+              onClose={handleCloseUserMenu}
+            >
+              {settings.map((setting) => (
+                <MenuItem
+                  key={setting}
+                  onClick={
+                    setting === "Logout" ? handleLogout : handleCloseUserMenu
+                  }
+                >
+                  <Typography sx={{ textAlign: "center" }}>
+                    {setting}
+                  </Typography>
+                </MenuItem>
+              ))}
+            </Menu>
+          </>
+        ) : (
+          <Button
+            component={Link}
+            to="sign-in"
+            color="inherit"
+            sx={{ order: { xs: 3, md: 0 } }}
+            // onClick={handleAuth}
+          >
+            Sign in
+          </Button>
+        )}
+      </Toolbar>
+    </AppBar>
+  );
+}
+export default Navbar;
